requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.8
django-allauth==65.13.1
django-crispy-forms==2.5
django-widget-tweaks==1.5.0
idna==3.11
kombu==5.6.2
packaging==26.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
pypdfium2==5.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
redis==7.3.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.6.0
